buncha changes: save and publish, invalid author, logged in comments, custom template tag "has group", post_detail restyle, draftlist restyle NOTE: custom 403 handler needs fix

+++ blog/views.py
@@ -3,6 +3,7 @@ from django.contrib.auth import login, authenticate
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.models import Group
+from django.contrib.auth.mixins import PermissionRequiredMixin
 from django.utils import timezone
 from django.views.generic import ListView, DetailView, CreateView
 from .models import Post, Comment, Category, CV
@@ -27,19 +28,21 @@ def post_detail(request, pk):
 @login_required
 @allowed_users(allowed_roles=['admin', 'author'])
 def post_new(request):
+    category_menu = Category.objects.all()
     if request.method == "POST":
         form = PostForm(request.POST)
         if form.is_valid():
-            post = form.save(commit=False) # Default commit is true, but in this case we want to add author THEN save!
+            post = form.save(commit=False) # Commit false so can add author before save
             post.author = request.user
-            # post.published_date = timezone.now()          # removed to separate drafting and publishing
+            if 'save_and_publish' in request.POST:
+                post.published_date = timezone.now()
             post.save()
             return redirect('post_detail', pk=post.pk)
     else:
         form = PostForm()
-    return render(request, 'blog/post_edit.html', {'form': form})
+    return render(request, 'blog/post_edit.html', {'form': form, 'category_menu': category_menu})
 
-# Alternative postnew view! NOTE: NOT IN USE
+# Alternative postnew view! NOTE: NOT IN USE (also should learn about mixins)
 class AddPostView(CreateView):
     model = Post
     form_class = PostForm
@@ -52,26 +55,28 @@ class AddPostView(CreateView):
 @allowed_users(allowed_roles=['admin', 'author'])
 def post_edit(request, pk):
     post = get_object_or_404(Post, pk=pk)
+    category_menu = Category.objects.all()
     if request.user == post.author or request.user.groups.filter(name='admin').exists():
         if request.method == "POST":
             form = PostForm(request.POST, instance=post)
             if form.is_valid():
                 post = form.save(commit=False)
                 post.author = request.user
-                # post.published_date = timezone.now()          # removed to separate drafting and publishing 
+                if 'save_and_publish' in request.POST:
+                    post.published_date = timezone.now()
                 post.save()
                 return redirect('post_detail', pk=post.pk)
         else:
             form = PostForm(instance=post)
-        return render(request, 'blog/post_edit.html', {'form': form})
+        return render(request, 'blog/post_edit.html', {'form': form, 'post': post, 'category_menu': category_menu})
     else:
         return redirect('invalid_author', pk=post.pk)
 
 # Show unpublished posts
 @login_required
-@allowed_users(allowed_roles=['admin'])
+@allowed_users(allowed_roles=['admin', 'author'])
 def post_draft_list(request):
-    posts = Post.objects.filter(published_date__isnull=True).order_by('created_date')
+    posts = Post.objects.filter(published_date__isnull=True).order_by('-created_date')
     return render(request, 'blog/post_draft_list.html', {'posts': posts})
 
 # Publish draft post
@@ -79,16 +84,22 @@ def post_draft_list(request):
 @allowed_users(allowed_roles=['admin', 'author'])
 def post_publish(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    post.publish()
-    return redirect('post_detail', pk=pk)
+    if request.user == post.author or request.user.groups.filter(name='admin').exists():
+        post.publish()
+        return redirect('post_detail', pk=pk)
+    else:
+        return redirect('invalid_author', pk=post.pk)
 
 # Delete existing post
 @login_required
 @allowed_users(allowed_roles=['admin', 'author'])
 def post_remove(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    post.delete()                           # Model.delete is standard django method.
-    return redirect('post_list')
+    if request.user == post.author or request.user.groups.filter(name='admin').exists():
+        post.delete()                           # Model.delete is standard django method.
+        return redirect('post_list')
+    else:
+        return redirect('invalid_author', pk=post.pk)
 
 def post_invalid_author(request, pk):
     post = get_object_or_404(Post, pk=pk)
@@ -102,6 +113,8 @@ def add_comment_to_post(request, pk):
         if form.is_valid():
             comment = form.save(commit=False)
             comment.post = post
+            if request.user.is_authenticated:
+                comment.approve()
             comment.save()
             return redirect('post_detail', pk=post.pk)
     else:
@@ -123,7 +136,8 @@ def comment_remove(request, pk):
     return redirect('post_detail', pk=comment.post.pk)
 
 # Add category NOTE: TRYING OUT CLASS VIEWS
-class AddCategoryView(CreateView):
+class AddCategoryView(PermissionRequiredMixin, CreateView): # We have to include mixin arg instead of equivalent decorator.
+    permission_required = 'blog.add_category'
     model = Category
     template_name = 'blog/add_category.html'
     fields = '__all__'
@@ -131,8 +145,11 @@ class AddCategoryView(CreateView):
 def category_view(request, category):
     category_posts = Post.objects.filter(category__iexact=category.replace('-', ' '))
     category_menu = Category.objects.all()
-    return render(request, 'blog/categories.html', {'category': category.title().replace('-', ' '), 'category_posts':category_posts, 'category_menu': category_menu})
+    return render(request, 'blog/categories.html', {'category': category.title().replace('-', ' '), 'category_posts': category_posts, 'category_menu': category_menu})
 
 def cv(request):
     form = CVForm()
     return render(request, 'blog/cv.html', {'form': form})
+
+def handler403(request, exception):
+    return render(request, '403.html')
